+++ askanna/__init__.py
@@ -1,6 +1,6 @@
 __author__ = "AskAnna Team"
 __email__ = "support@askanna.io"
-__version__ = "0.21.0"
+__version__ = "0.22.0"
 
 import re
 import sys