+++ tests/fixtures/responses/api/project.py
@@ -130,6 +130,23 @@ def project_response(
             matchers.json_params_matcher(
                 {
                     "workspace_suuid": "1234-1234-1234-1234",
+                    "name": "a new project",
+                    "description": "",
+                    "visibility": "PUBLIC",
+                }
+            )
+        ],
+        status=201,
+        content_type="application/json",
+        json=project_new_detail,
+    )
+    api_responses.add(
+        "POST",
+        url=askanna_url.project.project(),
+        match=[
+            matchers.json_params_matcher(
+                {
+                    "workspace_suuid": "1234-1234-1234-1234",
                     "name": "new-project",
                     "description": "",
                     "visibility": "PRIVATE",